Reduce rerendering of TaskName component
